Allowed primary module location to be configured
This change allows the primary module (JS & LESS) to have a configured
location. This is possible my adding an entry to the `modules` array in
the yeoman config file for the primary module and also having a
`primaryModule` key with the name of the primary module.
Also, modules no longer have to be in subdirectories of their parent
modules. When adding a new module, the parent directories will be
searched for a module defining file, however, if no modules are found,
the primary module will be attempted (regardless if the primary module
is in a parent directory of the new module).